requirements.txt
flask==2.0.1
Pillow==11.2.1
numpy==2.2.5
Werkzeug==2.0.1
--only-binary :all: 